Correção da página de cadastro
