+++ src/index.js
@@ -3,6 +3,7 @@ import { BrowserRouter, Routes, Route } from "react-router-dom";
 import Layout from "./pages/Layout";
 import Home from "./pages/Home";
 import NoPage from "./pages/NoPage";
+import Cadastro from "./pages/Cadastro/Cadastro";
 import 'bootstrap/dist/css/bootstrap.min.css';
 
 export default function App() {
@@ -12,6 +13,7 @@ export default function App() {
         <Route path="/" element={<Layout />}>
           <Route index element={<Home />} />
           <Route path="*" element={<NoPage />} />
+          <Route path="Cadastro" element={<Cadastro />} />
         </Route>
       </Routes>
     </BrowserRouter>
